Day 19
Introduced the 'ftax <player>' command, which sends a message and kicks the specified player after a delay. Updated DayXfeatures.txt to document the new command.

Made a @CmEvent, not complete yet
Made a Register.onChatMessageCancelable(), looks like it do replace at the correct emplacement and not before (e.g: if we replace a msg in /pl it will indeed appear in the right spot and not above the ------- as before)